Silence strict compiler warning

+++ source/blender/blenlib/BLI_hash.h
@@ -75,9 +75,9 @@ BLI_INLINE void BLI_hash_pointer_to_color(const void *ptr, int *r, int *g, int *
 	const size_t hash_b = BLI_hash_int((uint)((val & 0xffff0000) >> 32));
 	const size_t hash =
 	        hash_a ^ (hash_b + 0x9e3779b9 + (hash_a << 6) + (hash_a >> 2));
-	*r = (hash & 0xff0000) >> 16;
-	*g = (hash & 0x00ff00) >> 8;
-	*b = hash & 0x0000ff;
+	*r = (int)((hash & 0xff0000) >> 16);
+	*g = (int)((hash & 0x00ff00) >> 8);
+	*b = (int)(hash & 0x0000ff);
 }
 
 #endif // __BLI_HASH_H__
